Shifting auth related methods to its own file
